Optional subject_relation in GetRelationsRequest
When searching for relations, there's a difference between:

* Not filtering on subject_relation, in which case matching relations
are returned regardless of their subject_relation value.
* Filtering the results to only include relations that do not have a
subject relation.

The existing definitions do not offer a way to distinguish between
the two cases because the meaning of an empty string as the subject
relation can only be one of the two options.

This change makes subject_relation an optional field in which an
empty string is distinguishable from a null value.